Parallelization working, but getting getting wrong correlator, as it converges towards zero, not 0.6

+++ GluonicLQCD/correlators/correlator.h
@@ -9,14 +9,14 @@ protected:
 //    int m_N;
 //    int m_N_T;
     int *m_N;
-    int m_latticeSize;
+    double m_latticeSize;
 public:
 //    Correlator(int N, int N_T);
     Correlator();
     ~Correlator();
     virtual double calculate(Links *lattice);
     void setN(int *N);
-    void setLatticeSize(int latticeSize) { m_latticeSize = latticeSize; }
+    void setLatticeSize(int latticeSize) { m_latticeSize = double(latticeSize); }
 };
 
 #endif // CORRELATOR_H
